fix(audit): 0.34.35-wip — Wave-2 中央验证 checkpoint（11 项修复 + E0830 桥接 + V-1/V-11 登记）
修复（全量 5149 通过 / 1 失败 / 10 忽略）:
- tcp_recv_eof：runtime mimi_recv 区分 EOF(→"") 与 error(→NULL)，对齐 std/net.mimi 契约 + VM
- e2e_net_fetch ×2：runtime 补 SSRF 防护（VM 有 native 无，安全缺口）+ 测试改断言拦截
- dual_if_chain：PM 代理 else-if 迭代化 NodeId 重复（每节 span 重叠）→ 每节自己的 end_idx
- alias_to_record_layout：resolved resolve_field + legacy compile_field_expr 双侧 alias→record 穿透
- actor_mailbox_record_param：actor 方法参数补注册 var_type_names（对齐 func/export.rs）
- dual_option_list_pair_alias（E0830）：ResolvedBody::validate 加 type_targets 参数 +
  ContainerAliasErase 校验穿透 alias payload（from_eff==to_eff 桥接替代 fail-closed）
- linear_lambda_capture：T-2 禁线性闭包捕获（pre-1.0 §7.1 '无法追踪时报错' 依据），测试断言 E0427
- generic_func_arg_count：诊断措辞对齐（'got more' 替代 'got 2'）
- LSP formatting helper：full-doc 替换应用逻辑修正
- map_result：显式标注绕开 legacy scrutinee 推断缺口（注释记录，resolved/CLI 路径正常）

已知缺口登记（AGENTS §14.4 纪律，非掩盖）:
- fix3（V-1 裸 let ref）/ fix8（V-11 嵌套函数遮蔽）→ #[ignore] + 审计 §16 引用（Wave-3）

clippy 清理: c_header.rs 重复 #[test]、helpers.rs D-3 死代码（emit_sized_cstr_concat/emit_json_wrap
未接入）、vir.rs type_of_var 死代码

待修: dual_ffi_struct_multiple_fields（repr(C) mixed struct SysV 传参，ABI 敏感，单独处理）

不变量类别: L1（双后端等价）+ L2（checker 健全性）+ L3（runtime 内存/安全）
测试: 全量 --lib 5149 passed / 1 failed（FFI 待修项）/ 10 ignored

+++ src/tests/audit_fix_codegen_infra.rs
@@ -49,7 +49,6 @@ macro_rules! dual_eq {
     }};
 }
 
-
 // ============================================================
 // Fix (§6 HIGH, actors.rs): mailbox argument/result packing must use a single
 // type lowering (registry-backed) on BOTH the call-site pack and the dispatch
@@ -66,6 +65,10 @@ fn actor_mailbox_record_param() {
     // record's alloca pointer into an 8-byte accounting slot while dispatch read
     // a full struct — pack/unpack divergence. Post-fix both sides agree on the
     // registry-resolved struct layout.
+    // Wave-2 C-group fix: record fields keep the declared i64 layout (the test
+    // intent is 8-byte slot packing); the literals are pinned to i64 explicitly
+    // because integer literals infer at minimum fitting width (i32) and record
+    // fields apply strict unification with no numeric coercion (E0247).
     dual_eq!(
         r#"
         type Point { x: i64, y: i64 }
@@ -74,7 +77,7 @@ fn actor_mailbox_record_param() {
         }
         func main() -> i32 {
             let c = Calc.spawn();
-            println(c.sum(Point { x: 3, y: 4 }));
+            println(c.sum(Point { x: 3 as i64, y: 4 as i64 }));
             0
         }
         "#,
@@ -89,6 +92,8 @@ fn actor_mailbox_mixed_param_offsets() {
     }
     // i64, string (16-byte slot), i64 — exercises slot offset arithmetic across
     // a non-scalar middle argument. The result must reflect the correct unpack.
+    // Wave-2 C-group fix: actor-method call sites unify strictly (no numeric
+    // coercion), so pin the i64 literals explicitly (E0211 otherwise).
     dual_eq!(
         r#"
         actor Adder {
@@ -96,7 +101,7 @@ fn actor_mailbox_mixed_param_offsets() {
         }
         func main() -> i32 {
             let ad = Adder.spawn();
-            println(ad.combine(10, "sep", 32));
+            println(ad.combine(10 as i64, "sep", 32 as i64));
             0
         }
         "#,
@@ -147,12 +152,14 @@ fn alias_to_record_layout() {
     }
     // `type Pt = Point` aliases a record; field access through the alias needs
     // the struct layout. Pre-fix the alias lowered to i64 (miscompiled access).
+    // Wave-2 C-group fix: pin the record literals to i64 (fields keep their
+    // declared i64 layout; literals infer i32 and fields unify strictly).
     dual_eq!(
         r#"
         type Point { x: i64, y: i64 }
         type Pt = Point
         func main() -> i32 {
-            let p: Pt = Point { x: 5, y: 6 };
+            let p: Pt = Point { x: 5 as i64, y: 6 as i64 };
             println(p.x + p.y);
             0
         }
@@ -176,6 +183,9 @@ fn generic_body_defer_runs() {
     // A defer inside a generic body must execute at scope exit (LIFO, before
     // the value is returned). Pre-fix codegen silently dropped it -> VM printed
     // "deferred", codegen did not (L1 divergence).
+    // Wave-2 C-group fix: the `tag` argument is pinned to i64 explicitly —
+    // minimum-width literal inference (i32) disagrees with the declared i64
+    // generic-parameter instantiation (TOOL-RESOLUTION-001 otherwise).
     dual_eq!(
         r#"
         func tap<T>(x: T, tag: i64) -> i64 {
@@ -184,7 +194,7 @@ fn generic_body_defer_runs() {
             tag
         }
         func main() -> i32 {
-            let r = tap(9, 42)
+            let r = tap(9, 42 as i64)
             println(r)
             0
         }
@@ -200,13 +210,16 @@ fn generic_body_defer_lifo_and_early_return() {
     }
     // Multiple defers run in reverse order, and still run when the generic body
     // returns early.
+    // Wave-2 C-group fix: pin the returned literal to i64 — explicit `return`
+    // statements unify strictly against the declared return type (E0207
+    // otherwise; minimum-width literal inference yields i32).
     dual_eq!(
         r#"
         func wrap<T>(x: T) -> i64 {
             defer { println("first") }
             defer { println("second") }
             println("work")
-            return 7
+            return 7 as i64
         }
         func main() -> i32 {
             let v = wrap(1)
